+++ packages/create-pantheon-decoupled-kit/__tests__/index.test.ts
@@ -1,19 +1,14 @@
 // import like this so we can vi.spyOn
 import * as bin from '../src/index';
+import * as utils from '../src/utils/index';
 const { parseArgs, main } = bin;
-import { decoupledKitTestGenerators } from './testGenerators/generators/index';
-import inquirer from 'inquirer';
 import chalk from 'chalk';
-import * as nodePlop from 'node-plop';
-import { helpMenu } from '../src/utils/helpMenu';
+import inquirer from 'inquirer';
+import { decoupledKitGenerators } from '../src/generators';
+
 import pkg from '../package.json' assert { type: 'json' };
 
-vi.mock('node-plop');
 vi.mock('inquirer');
-
-afterEach(() => {
-	vi.restoreAllMocks();
-});
 
 describe('parseArgs()', () => {
 	it('should parse args from the command line', () => {
@@ -45,49 +40,82 @@ describe('parseArgs()', () => {
 	});
 });
 
-describe('setGenerators()', () => {
-	it('should set all generators passed into it', async () => {
-		const setGeneratorSpy = vi.spyOn(bin, 'setGenerators');
-		const plop = await bin.setGenerators(decoupledKitTestGenerators);
-
-		expect(setGeneratorSpy).toHaveBeenCalledOnce();
-		expect(plop.setGenerator).toHaveBeenCalledTimes(2);
-		expect(plop.setActionType).toHaveBeenCalledTimes(3);
-		expect(plop).toHaveProperty('setPlopfilePath');
-		expect(plop).toHaveProperty('getGenerator');
-		expect(plop.getGeneratorList()).toEqual([
-			{ name: 'test-add' },
-			{ name: 'test-append' },
-		]);
-	});
-});
-
 describe('main()', () => {
 	beforeEach((context) => {
 		context.parseSpy = vi.spyOn(bin, 'parseArgs');
 		context.mainSpy = vi.spyOn(bin, 'main');
+		context.actionRunnerSpy = vi.spyOn(utils, 'actionRunner');
 		context.promptSpy = vi.spyOn(inquirer, 'prompt');
 		context.logSpy = vi.spyOn(console, 'log');
 		context.errorSpy = vi.spyOn(console, 'error');
-		context.plopSpy = vi.spyOn(nodePlop, 'default');
+
+		context.actionRunnerSpy.mockImplementation(vi.fn());
 	});
+	afterEach(() => {
+		vi.restoreAllMocks();
+	});
+
+	it('should show the help menu if --help or -h is in the args', async ({
+		logSpy,
+	}) => {
+		// --help
+		process.argv = ['node', 'bin.js', '--help'];
+
+		await main(parseArgs(), decoupledKitGenerators);
+
+		expect(logSpy).toHaveBeenLastCalledWith(
+			utils.helpMenu(decoupledKitGenerators),
+		);
+
+		// -h
+		process.argv = ['node', 'bin.js', '-h'];
+
+		await main(parseArgs(), decoupledKitGenerators);
+
+		expect(logSpy).toHaveBeenLastCalledWith(
+			utils.helpMenu(decoupledKitGenerators),
+		);
+	});
+
+	it('should show the version if --version or -v is in the args', async ({
+		logSpy,
+	}) => {
+		// --version
+		process.argv = ['node', 'bin.js', '--version'];
+
+		await main(parseArgs(), decoupledKitGenerators);
+
+		expect(logSpy).toHaveBeenLastCalledWith(`v${pkg.version}`);
+
+		// -v
+		process.argv = ['node', 'bin.js', '-v'];
+
+		await main(parseArgs(), decoupledKitGenerators);
+
+		expect(logSpy).toHaveBeenLastCalledWith(`v${pkg.version}`);
+	});
+
 	it('should accept any number of positional arguments as generators to run', async ({
 		parseSpy,
 		mainSpy,
 		promptSpy,
 	}) => {
-		process.argv = ['node', 'bin.js', 'test-add', 'test-append'];
+		process.argv = ['node', 'bin.js', 'next-drupal', 'next-drupal-umami-addon'];
+		promptSpy.mockImplementationOnce(() => ({ outDir: 'test' }));
+		promptSpy.mockImplementationOnce(() => ({ appName: 'test' }));
 
 		await bin.main(
 			bin.parseArgs(process.argv.slice(2)),
-			decoupledKitTestGenerators,
+			decoupledKitGenerators,
 		);
 
 		expect(parseSpy).toBeCalled;
 		expect(parseSpy).toBeCalledWith(process.argv.slice(2));
 		expect(mainSpy).toHaveBeenLastCalledWith(
 			{
-				_: ['test-add', 'test-append'],
+				_: ['next-drupal', 'next-drupal-umami-addon'],
+				appName: 'test',
+				outDir: 'test',
 				force: false,
 				silent: false,
 				help: false,
@@ -95,7 +123,7 @@ describe('main()', () => {
 				version: false,
 				v: false,
 			},
-			decoupledKitTestGenerators,
+			decoupledKitGenerators,
 		);
 		expect(promptSpy).toHaveBeenCalledTimes(2);
 	});
@@ -106,147 +134,98 @@ describe('main()', () => {
 	}) => {
 		process.argv = ['node', 'bin.js', 'not-valid-generator-name'];
 
-		// inquirer is mocked so we know the type is wrong.
-		//@ts-ignore
-		inquirer.prompt.mockImplementationOnce(() => ({
-			generators: ['test-add'],
+		promptSpy.mockImplementationOnce(() => ({
+			generators: ['next-wp'],
 		}));
 
-		await main(parseArgs(), decoupledKitTestGenerators);
+		await bin.main(parseArgs(), decoupledKitGenerators);
 
-		expect(logSpy).toHaveBeenLastCalledWith(
+		expect(logSpy).toHaveBeenCalledWith(
 			chalk.yellow('No generator found with name not-valid-generator-name.'),
 		);
 		// We expect 2 calls, the first prompts with the list of available generators
 		// the second runs the generator prompts
 		expect(promptSpy).toHaveBeenCalledTimes(2);
 	});
-	it('should exit with a message if no valid generators are selected', async ({
-		logSpy,
-		errorSpy,
-		promptSpy,
-	}) => {
-		process.argv = ['node', 'bin.js', 'not-valid-generator-name'];
 
-		await main(parseArgs(), decoupledKitTestGenerators);
+	it.fails(
+		'should throw error if there is no valid generator selected to run',
+		async ({ mainSpy, promptSpy }) => {
+			process.argv = ['node', 'bin.js', 'not-valid-generator-name'];
+			promptSpy.mockImplementationOnce(() => ({
+				generators: [],
+			}));
+			await bin.main(parseArgs(), decoupledKitGenerators);
 
-		expect(logSpy).toHaveBeenLastCalledWith(
-			chalk.yellow('No generator found with name not-valid-generator-name.'),
-		);
-		expect(errorSpy).toHaveBeenLastCalledWith(
-			chalk.red(
+			expect(promptSpy).toHaveBeenCalledOnce();
+			expect(mainSpy).toThrowError(`
+			${chalk.red(
 				'No generators were selected. Use positional arguments or choose from the prompt.',
-			),
-		);
-		expect(promptSpy).toHaveBeenCalledTimes(1);
+			)}
+Valid generators: ${decoupledKitGenerators.map(({ name }) => name).join(', ')}
+To see this list at any time, use the --help command.`);
+		},
+	);
+
+	it.fails(
+		'should throw an error if message if no valid generators are selected',
+		async ({ promptSpy, mainSpy }) => {
+			process.argv = ['node', 'bin.js', 'not-valid-generator-name'];
+			promptSpy.mockImplementationOnce(() => ({
+				generators: [],
+			}));
+
+			await bin.main(parseArgs(), decoupledKitGenerators);
+
+			expect(mainSpy).toThrowErrorMatchingInlineSnapshot(
+				chalk.red('Could not find any generators to run.'),
+			);
+			expect(promptSpy).toHaveBeenCalledTimes(1);
+		},
+	);
+
+	it('should add arbitrary from generator.data to the args', async ({
+		actionRunnerSpy,
+	}) => {
+		// gatsby-wp has arbitrary data used to determine
+		// adding the pnpm plugin
+		process.argv = ['node', 'bin.js', 'gatsby-wp'];
+		const data = {
+			_: ['gatsby-wp'],
+			force: false,
+			gatsbyPnpmPlugin: true,
+			h: false,
+			help: false,
+			silent: false,
+			v: false,
+			version: false,
+		};
+
+		await main(parseArgs(), decoupledKitGenerators);
+
+		expect(
+			actionRunnerSpy.mock.lastCall && actionRunnerSpy.mock.lastCall[0].data,
+		).toEqual(data);
 	});
 
-	it('should log changes', async ({ logSpy, plopSpy }) => {
-		process.argv = ['node', 'bin.js', 'test-add'];
-		vi.mocked(nodePlop.default).mockReturnValueOnce({
-			//@ts-ignore
-			setPlopfilePath: vi.fn(),
-			getPlopfilePath: vi.fn(),
-			setPartial: vi.fn(),
-			getGenerator: vi.fn().mockImplementation(() => ({
-				runActions: vi.fn().mockReturnValue({
-					changes: [{ type: 'add', path: '/test' }],
-					failures: [],
-				}),
-			})),
-			setActionType: vi.fn(),
-			setGenerator: vi.fn(),
-			getGeneratorList: vi
-				.fn()
-				.mockImplementation(() => [
-					{ name: 'test-add' },
-					{ name: 'test-append' },
-				]),
-			setHelper: vi.fn(),
-		});
-
-		await main(parseArgs(), decoupledKitTestGenerators);
+	it('should log changes', async ({ logSpy, promptSpy }) => {
+		process.argv = ['node', 'bin.js', 'next-drupal'];
+		promptSpy.mockImplementationOnce(() => ({ outDir: 'test' }));
+		promptSpy.mockImplementationOnce(() => ({ appName: 'test' }));
+		await main(parseArgs(), decoupledKitGenerators);
 
 		expect(logSpy).toHaveBeenLastCalledWith(
-			chalk.green('add'),
-			chalk.cyan('/test'),
-		);
-		expect(plopSpy).toHaveBeenCalledTimes(1);
-	});
-
-	it('should log failures', async ({ errorSpy }) => {
-		vi.mocked(nodePlop.default).mockReturnValueOnce({
-			//@ts-ignore
-			setPlopfilePath: vi.fn(),
-			getPlopfilePath: vi.fn(),
-			setPartial: vi.fn(),
-			getGenerator: vi.fn().mockImplementation(() => ({
-				runActions: vi.fn().mockReturnValue({
-					changes: [],
-					failures: [
-						{
-							error: `File already exists\n  -> /test/test.js`,
-						},
-					],
-				}),
-			})),
-			setActionType: vi.fn(),
-			setGenerator: vi.fn(),
-			getGeneratorList: vi
-				.fn()
-				.mockImplementation(() => [
-					{ name: 'test-add' },
-					{ name: 'test-append' },
-				]),
-			setHelper: vi.fn(),
-		});
-		await main(parseArgs(), decoupledKitTestGenerators);
-		expect(errorSpy).toHaveBeenLastCalledWith(
-			chalk.red(`File already exists\n  -> /test/test.js`),
+			`${chalk.yellow('cd')} into ${chalk.bold.magenta(
+				'test',
+			)} to start developing!`,
 		);
 	});
 
 	it('should not console.log if "silent" arg is true', async ({ logSpy }) => {
-		process.argv = ['node', 'bin.js', 'test-add', '--silent'];
+		process.argv = ['node', 'bin.js', 'next-wp', '--silent'];
 
-		await main(parseArgs(), decoupledKitTestGenerators);
+		await main(parseArgs(), decoupledKitGenerators);
 
 		expect(logSpy).toHaveBeenCalledTimes(0);
-	});
-
-	it('should show the help menu if --help or -h is in the args', async ({
-		logSpy,
-	}) => {
-		// --help
-		process.argv = ['node', 'bin.js', '--help'];
-
-		await main(parseArgs(), decoupledKitTestGenerators);
-
-		expect(logSpy).toHaveBeenLastCalledWith(helpMenu);
-
-		// -h
-		process.argv = ['node', 'bin.js', '-h'];
-
-		await main(parseArgs(), decoupledKitTestGenerators);
-
-		expect(logSpy).toHaveBeenLastCalledWith(helpMenu);
-	});
-
-	it('should show the version if --version or -v is in the args', async ({
-		logSpy,
-	}) => {
-		// --version
-		process.argv = ['node', 'bin.js', '--version'];
-
-		await main(parseArgs(), decoupledKitTestGenerators);
-
-		expect(logSpy).toHaveBeenLastCalledWith(`v${pkg.version}`);
-
-		// -v
-		process.argv = ['node', 'bin.js', '-v'];
-
-		await main(parseArgs(), decoupledKitTestGenerators);
-
-		expect(logSpy).toHaveBeenLastCalledWith(`v${pkg.version}`);
 	});
 });